Fix hidden plot lines
